Move content of phenotypes page into a separate component
This is required in order to determine if the page is using dark mode.
